backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
boto3==1.36.7
python-dotenv==1.0.1
gunicorn==23.0.0
